post merge fix

+++ applications/services/gui/canvas.c
@@ -128,9 +128,9 @@ size_t canvas_current_font_height(const Canvas* canvas) {
     return font_height;
 }
 
-uint8_t canvas_current_font_width(const Canvas* canvas) {
+size_t canvas_current_font_width(const Canvas* canvas) {
     furi_assert(canvas);
-    return (uint8_t)u8g2_GetMaxCharWidth(&canvas->fb);
+    return (size_t)u8g2_GetMaxCharWidth(&canvas->fb);
 }
 
 const CanvasFontParameters* canvas_get_font_params(const Canvas* canvas, Font font) {
